Porting all tests from disposable mixin to disposable util
