+++ tests/unit/utils/disposable-test.js
@@ -1,4 +1,5 @@
 import EmberObject from '@ember/object';
+import { run } from '@ember/runloop';
 import { module, test } from 'qunit';
 import {
   registerDisposable,
@@ -11,7 +12,7 @@ module('ember-lifeline/utils/disposable', {
   },
 
   afterEach() {
-    this.subject = null;
+    run(this.subject, 'destroy');
   },
 });
 
@@ -115,4 +116,82 @@ test('registerDisposable sets up willDestroy only once', function(assert) {
   registerDisposable(this.subject, dispose);
 
   assert.notOk(this.subject.willDestroy.twice, 'willDestroy only patched once');
+});
+
+test('disposable invoked explicitly disposes of disposable', function(assert) {
+  assert.expect(2);
+
+  let callCount = 0;
+  let dispose = () => {
+    callCount++;
+  };
+
+  let disposable = registerDisposable(this.subject, dispose);
+
+  disposable.dispose();
+
+  assert.equal(callCount, 1, 'disposable is called');
+  assert.ok(
+    this.subject._registeredDisposables[0].disposed,
+    'disposable marked as disposed'
+  );
+});
+
+test('disposable invoked explicitly multiple times is only invoked once', function(assert) {
+  assert.expect(2);
+
+  let callCount = 0;
+  let dispose = () => {
+    callCount++;
+  };
+
+  let disposable = registerDisposable(this.subject, dispose);
+
+  disposable.dispose();
+  disposable.dispose();
+
+  assert.equal(callCount, 1, 'disposable is called');
+  assert.ok(
+    this.subject._registeredDisposables[0].disposed,
+    'disposable marked as disposed'
+  );
+});
+
+test('runDisposables: runs all disposables when destroying', function(assert) {
+  assert.expect(2);
+
+  let dispose = () => {};
+  let disposeTheSecond = () => {};
+
+  registerDisposable(this.subject, dispose);
+  registerDisposable(this.subject, disposeTheSecond);
+
+  assert.equal(
+    this.subject._registeredDisposables.length,
+    2,
+    'two disposables are registered'
+  );
+
+  run(this.subject, 'destroy');
+
+  assert.equal(
+    this.subject._registeredDisposables.length,
+    0,
+    'no disposables are registered'
+  );
+});
+
+test('runDisposables: sets all disposables to disposed', function(assert) {
+  assert.expect(2);
+
+  let dispose = () => {};
+  let disposeTheSecond = () => {};
+
+  let disposable = registerDisposable(this.subject, dispose);
+  let disposableTheSecond = registerDisposable(this.subject, disposeTheSecond);
+
+  run(this.subject, 'destroy');
+
+  assert.ok(disposable.disposed, 'first disposable is desposed');
+  assert.ok(disposableTheSecond.disposed, 'second disposable is desposed');
 });
